feat: slug 네임 변경 및 GET 요청 위치 변경
